+++ src/hb-aat-layout-morx-table.hh
@@ -1033,7 +1033,7 @@ struct ChainSubtable
   template <typename T>
   friend struct Chain;
 
-  unsigned int get_size () const     { return length; }
+  size_t get_size () const     { return length; }
   unsigned int get_type () const     { return coverage & 0xFF; }
   unsigned int get_coverage () const { return coverage >> (sizeof (HBUINT) * 8 - 8); }
 
@@ -1243,7 +1243,7 @@ struct Chain
       c->reverse_buffer ();
   }
 
-  unsigned int get_size () const { return length; }
+  size_t get_size () const { return length; }
 
   template <typename context_t, typename ...Ts>
   typename context_t::return_t dispatch (context_t *c, Ts&&... ds) const
